refactor(core-dotnet-domain): Updated version to use ulong type

+++ libs/core/dotnet/application/Attributes/CommandVersionAttribute.cs
@@ -5,7 +5,7 @@ namespace OpenSystem.Core.Application.Attributes
     [AttributeUsage(AttributeTargets.Class)]
     public class CommandVersionAttribute : VersionAttribute
     {
-        public CommandVersionAttribute(string name, uint version)
+        public CommandVersionAttribute(string name, ulong version)
             : base(name, version) { }
     }
 }
